Castor3D: Moved LightingModelFactory to its own file.
