+++ include/Core/Castor3D/Shader/Shaders/GlslLighting.hpp
@@ -18,205 +18,6 @@ See LICENSE file in root folder
 
 namespace castor3d::shader
 {
-	static const castor::String TOO_MANY_OBJECTS = cuT( "Too many lighting models declared" );
-	static const castor::String DUPLICATE_OBJECT = cuT( "Lighting model already declared" );
-
-	class LightingModelFactory
-	{
-	public:
-		using Obj = LightingModel;
-		using Key = castor::String;
-		using PtrType = LightingModelUPtr;
-		using Creator = LightingModelCreator;
-		using ObjPtr = PtrType;
-		struct Entry
-		{
-			Key name;
-			Creator create;
-			LightingModelID lightingModelId;
-			BackgroundModelID backgroundModelId;
-		};
-		using ObjCont = std::vector< Entry >;
-
-	public:
-		/**
-		 *\~english
-		 *\brief		Registers an object type
-		 *\param[in]	key		The object type
-		 *\param[in]	creator	The object creation function
-		 *\~french
-		 *\brief		Enregistre un type d'objet
-		 *\param[in]	key		Le type d'objet
-		 *\param[in]	creator	La fonction de création d'objet
-		 */
-		LightingModelID registerType( Key const & key
-			, BackgroundModelID backgroundModelId
-			, Creator create )
-		{
-			LightingModelID lightingModelId{};
-			auto it = std::find_if( m_registered.begin()
-				, m_registered.end()
-				, [&key]( Entry const & lookup )
-				{
-					return key == lookup.name;
-				} );
-
-			if ( it == m_registered.end() )
-			{
-				if ( m_currentId > MaxLightingModels )
-				{
-					CU_Exception( TOO_MANY_OBJECTS );
-				}
-
-				lightingModelId = ++m_currentId;
-			}
-			else
-			{
-				lightingModelId = it->lightingModelId;
-			}
-
-			auto & entry = m_registered.emplace_back();
-			entry.name = key;
-			entry.create = create;
-			entry.lightingModelId = lightingModelId;
-			entry.backgroundModelId = backgroundModelId;
-			return lightingModelId;
-		}
-		/**
-		 *\~english
-		 *\brief		Unregisters an object type.
-		 *\param[in]	key					The object type.
-		 *\param[in]	backgroundModelId	The background model ID.
-		 *\~french
-		 *\brief		Désenregistre un type d'objet.
-		 *\param[in]	key					Le type d'objet.
-		 *\param[in]	backgroundModelId	L'ID du modèle de fond.
-		 */
-		void unregisterType( Key const & key
-			, BackgroundModelID backgroundModelId )
-		{
-			auto it = std::find_if( m_registered.begin()
-				, m_registered.end()
-				, [&key, backgroundModelId]( Entry const & lookup )
-				{
-					return key == lookup.name
-						&& backgroundModelId == lookup.backgroundModelId;
-				} );
-
-			if ( it != m_registered.end() )
-			{
-				m_registered.erase( it );
-			}
-		}
-		/**
-		 *\~english
-		 *\brief		Unregisters an object type.
-		 *\param[in]	lightingModelId		The lighting model ID.
-		 *\param[in]	backgroundModelId	The background model ID.
-		 *\~french
-		 *\brief		Désenregistre un type d'objet.
-		 *\param[in]	lightingModelId		L'ID du modèle d'éclairage.
-		 *\param[in]	backgroundModelId	L'ID du modèle de fond.
-		 */
-		void unregisterType( LightingModelID lightingModelId
-			, BackgroundModelID backgroundModelId )
-		{
-			auto it = std::find_if( m_registered.begin()
-				, m_registered.end()
-				, [lightingModelId, backgroundModelId]( Entry const & lookup )
-				{
-					return lightingModelId == lookup.lightingModelId
-					&& backgroundModelId == lookup.backgroundModelId;
-				} );
-
-			if ( it != m_registered.end() )
-			{
-				m_registered.erase( it );
-			}
-		}
-		/**
-		 *\~english
-		 *\return		The unique lighting models IDs.
-		 *\~french
-		 *\return		Les ID uniques des modèles d'éclairage.
-		 */
-		std::vector< LightingModelID > getLightingModelsID()const
-		{
-			std::unordered_set< LightingModelID > result;
-
-			for ( auto & entry : m_registered )
-			{
-				result.emplace( entry.lightingModelId );
-			}
-
-			return { result.begin(), result.end() };
-		}
-		/**
-		 *\~english
-		 *\param[in]	backgroundModelId	The background model ID.
-		 *\return		The lighting models for given background model.
-		 *\~french
-		 *\param[in]	backgroundModelId	L'ID du modèle de fond.
-		 *\return		Les modèles d'éclairage du mod_le de fond donné.
-		 */
-		std::vector< LightingModelID > getLightingModelsID( BackgroundModelID backgroundModelId )const
-		{
-			std::vector< LightingModelID > result;
-
-			for ( auto & entry : m_registered )
-			{
-				if ( entry.backgroundModelId == backgroundModelId )
-				{
-					result.push_back( entry.lightingModelId );
-				}
-			}
-
-			return result;
-		}
-		/**
-		 *\~english
-		 *\brief		Creates an object from an ID.
-		 *\param[in]	id		The object ID.
-		 *\param[in]	params	The creation parameters.
-		 *\return		The created object.
-		 *\~french
-		 *\brief		Crée un objet à partir d'un ID.
-		 *\param[in]	id		L'ID d'objet.
-		 *\param[in]	params	Les paramètres de création.
-		 *\return		L'objet créé.
-		 */
-		template< typename ... Parameters >
-		ObjPtr create( LightingModelID lightingModelId
-			, BackgroundModelID backgroundModelId
-			, Parameters && ... params )const
-		{
-			if ( lightingModelId == 0u )
-			{
-				return nullptr;
-			}
-
-			auto it = std::find_if( m_registered.begin()
-				, m_registered.end()
-				, [lightingModelId, backgroundModelId]( Entry const & lookup )
-				{
-					return lightingModelId == lookup.lightingModelId
-						&& backgroundModelId == lookup.backgroundModelId;
-				} );
-
-			if ( it == m_registered.end() )
-			{
-				CU_Exception( castor::ERROR_UNKNOWN_OBJECT );
-			}
-
-			return it->create( lightingModelId
-				, std::forward< Parameters >( params )... );
-		}
-
-	private:
-		LightingModelID m_currentId{};
-		ObjCont m_registered;
-	};
-
 	class LightingModel
 	{
 	public:
