+++ source/Core/Castor3D/Material/Pass/Component/Lighting/LightingModelComponent.cpp
@@ -7,6 +7,7 @@
 #include "Castor3D/Material/Pass/Component/PassComponentRegister.hpp"
 #include "Castor3D/Miscellaneous/Logger.hpp"
 #include "Castor3D/Scene/SceneFileParser.hpp"
+#include "Castor3D/Shader/LightingModelFactory.hpp"
 #include "Castor3D/Shader/ShaderBuffers/PassBuffer.hpp"
 #include "Castor3D/Shader/Shaders/GlslBlendComponents.hpp"
 #include "Castor3D/Shader/Shaders/GlslLighting.hpp"
