+++ source/Core/Castor3D/Render/Opaque/DeferredRendering.cpp
@@ -15,8 +15,8 @@
 #include "Castor3D/Scene/Camera.hpp"
 #include "Castor3D/Scene/Scene.hpp"
 #include "Castor3D/Scene/Background/Background.hpp"
+#include "Castor3D/Shader/LightingModelFactory.hpp"
 #include "Castor3D/Shader/ShaderBuffers/PassBuffer.hpp"
-#include "Castor3D/Shader/Shaders/GlslLighting.hpp"
 
 #include <CastorUtils/Design/ResourceCache.hpp>
