+++ source/Core/Castor3D/Shader/Shaders/GlslLighting.cpp
@@ -19,7 +19,6 @@
 #include <ShaderAST/Expr/ExprComma.hpp>
 #include <ShaderWriter/Source.hpp>
 
-CU_ImplementCUSmartPtr( castor3d::shader, LightingModelFactory )
 CU_ImplementCUSmartPtr( castor3d::shader, LightingModel )
 
 namespace castor3d::shader
